+++ iroh-dns/src/lib.rs
@@ -1,6 +1,4 @@
 pub mod discovery;
-pub mod dns;
 pub mod packet;
 pub mod publish;
 pub mod resolve;
-// pub mod sig0;
